Fixes to MPC comet import.  Moved SSObjectType <-> string mapping to SSObject.cpp and created inverse mapping.  Moved SSStar::setNames()/getNames() into SSObject base class so SSPlanet can use it too.  SSPlanets now get a single SSIdentifier.  strtofloat64() so it actually reads a double.  Comments in SSOrbit.

+++ SSUtilities.cpp
@@ -121,7 +121,7 @@ double strtofloat64 ( string str )
     
     try
     {
-        d = stof ( str );
+        d = stod ( str );
     }
     catch ( ... )
     {
